hardening: address sprint 2.0 qa fixes
